+++ source/fab/dep_tree.py
@@ -10,16 +10,90 @@ Classes and helper functions related to the dependency tree, as created by the a
 
 # todo: we've since adopted the term "source tree", so we should probably rename this module to match.
 import logging
+from abc import ABC
 from pathlib import Path
-from typing import Set, Dict, Iterable, List
+from typing import Set, Dict, Iterable, List, Union, Optional, Any
 
-from fab.parse.c import AnalysedC
-
-from fab.parse.fortran import AnalysedFortran
-
-from fab.parse import AnalysedDependent
+from fab.parse import AnalysedFile
 
 logger = logging.getLogger(__name__)
+
+
+# Todo: Better name? It's an analysed file in a dependency tree
+#       (as opposed to an analysed x90 for example, which isn't part of this tree dependency analysis).
+class AnalysedDependent(AnalysedFile, ABC):
+    """
+    An :class:`~fab.parse.AnalysedFile` which can depend on others, and be a dependency.
+    Instances of this class are nodes in a source dependency tree.
+
+    During parsing, the symbol definitions and dependencies are filled in.
+    During dependency analysis, symbol dependencies are turned into file dependencies.
+
+    """
+    def __init__(self, fpath: Union[str, Path], file_hash: Optional[int] = None,
+                 symbol_defs: Optional[Iterable[str]] = None, symbol_deps: Optional[Iterable[str]] = None,
+                 file_deps: Optional[Iterable[Path]] = None):
+        """
+        :param fpath:
+            The source file that was analysed.
+        :param file_hash:
+            The hash of the source. If omitted, Fab will evaluate lazily.
+        :param symbol_defs:
+            Set of symbol names defined by this source file.
+        :param symbol_deps:
+            Set of symbol names used by this source file.
+            Can include symbols in the same file.
+        :param file_deps:
+            Other files on which this source depends. Must not include itself.
+            This attribute is calculated during symbol analysis, after everything has been parsed.
+
+        """
+        super().__init__(fpath=fpath, file_hash=file_hash)
+
+        self.symbol_defs: Set[str] = set(symbol_defs or {})
+        self.symbol_deps: Set[str] = set(symbol_deps or {})
+        self.file_deps: Set[Path] = set(file_deps or [])
+
+        assert all([d and len(d) for d in self.symbol_defs]), "bad symbol definitions"
+        assert all([d and len(d) for d in self.symbol_deps]), "bad symbol dependencies"
+
+    def add_symbol_def(self, name):
+        assert name and len(name)
+        self.symbol_defs.add(name.lower())
+
+    def add_symbol_dep(self, name):
+        assert name and len(name)
+        self.symbol_deps.add(name.lower())
+
+    def add_file_dep(self, name):
+        self.file_deps.add(Path(name))
+
+    @classmethod
+    def field_names(cls):
+        return super().field_names() + [
+            'symbol_defs',
+            'symbol_deps',
+            'file_deps',
+        ]
+
+    def to_dict(self) -> Dict[str, Any]:
+        result = super().to_dict()
+        result.update({
+            "symbol_defs": list(sorted(self.symbol_defs)),
+            "symbol_deps": list(sorted(self.symbol_deps)),
+            "file_deps": list(sorted(map(str, self.file_deps))),
+        })
+        return result
+
+    @classmethod
+    def from_dict(cls, d):
+        return cls(
+            fpath=Path(d["fpath"]),
+            file_hash=d["file_hash"],
+            symbol_defs=set(d["symbol_defs"]),
+            symbol_deps=set(d["symbol_deps"]),
+            file_deps=set(map(Path, d["file_deps"])),
+        )
 
 
 def extract_sub_tree(source_tree: Dict[Path, AnalysedDependent], root: Path, verbose=False)\
@@ -73,28 +147,6 @@ def _extract_sub_tree(src_tree: Dict[Path, AnalysedDependent], key: Path,
         # add this child dep
         _extract_sub_tree(
             src_tree=src_tree, key=file_dep, dst_tree=dst_tree, missing=missing, verbose=verbose, indent=indent + 1)
-
-
-def add_mo_commented_file_deps(source_tree: Dict[Path, AnalysedDependent]):
-    """
-    Handle dependencies from Met Office "DEPENDS ON:" code comments which refer to a c file.
-    These are the comments which refer to a .o file and not those which just refer to symbols.
-
-    :param source_tree:
-        The source tree of analysed files.
-
-    """
-    # todo: this would be better if filtered by type, i,e, AnalysedFortran & AnalysedC
-    analysed_fortran: List[AnalysedFortran] = filter_source_tree(source_tree, '.f90')  # type: ignore
-    analysed_c: List[AnalysedC] = filter_source_tree(source_tree, '.c')  # type: ignore
-
-    lookup = {c.fpath.name: c for c in analysed_c}
-    num_found = 0
-    for f in analysed_fortran:
-        num_found += len(f.mo_commented_file_deps)
-        for dep in f.mo_commented_file_deps:
-            f.file_deps.add(lookup[dep].fpath)
-    logger.info(f"processed {num_found} DEPENDS ON file dependencies")
 
 
 def filter_source_tree(source_tree: Dict[Path, AnalysedDependent], suffixes: Iterable[str]) -> List[AnalysedDependent]:
